Reorder refinery directory
