+++ lib/refinery/assets.rb
