+++ lib/refinery/operations/delete.rb
