+++ test.rb
@@ -1,0 +1,6 @@
+require "refinery"
+
+Refinery::Container.finalize!
+
+require "irb"
+IRB.start
